feat: base64 move to conversion package
